+++ src/routes/Licenses.js
@@ -5,8 +5,9 @@ import { SearchAndSort } from '@folio/stripes/smart-components';
 import getSASParams from '../util/getSASParams';
 import packageInfo from '../../package';
 
-import ViewLicense from '../components/ViewLicense';
 import EditLicense from '../components/EditLicense';
+import LicenseFilters from '../components/LicenseFilters';
+import ViewLicense from '../components/ViewLicense';
 
 const INITIAL_RESULT_COUNT = 100;
 
@@ -18,21 +19,59 @@ export default class Licenses extends React.Component {
       path: 'licenses/licenses',
       params: getSASParams({
         searchKey: 'name',
+        columnMap: {
+          'Name': 'name',
+          'Description': 'description',
+        }
       })
     },
-    query: {},
+    statusValues: {
+      type: 'okapi',
+      path: 'licenses/refdata/License/status',
+    },
+    query: { initialValue: {} },
     resultCount: { initialValue: INITIAL_RESULT_COUNT },
+    selectedLicenseId: { initialValue: '' },
   });
 
   static propTypes = {
     resources: PropTypes.shape({
-      records: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.object), PropTypes.object]),
+      query: PropTypes.object,
+      records: PropTypes.object,
+      statusValues: PropTypes.object,
     }),
     mutator: PropTypes.object,
     onSelectRow: PropTypes.func,
+    browseOnly: PropTypes.bool,
   };
 
-  create = (license) => {
+  state = {
+    activeFilters: [],
+  }
+
+  handleFilterChange = ({ name, values }) => {
+    this.setState((prevState) => ({
+      activeFilters: {
+        ...prevState.activeFilters,
+        [name]: values,
+      }
+    }), () => {
+      const { activeFilters } = this.state;
+
+      const filters = Object.keys(activeFilters)
+        .map((filterName) => {
+          return activeFilters[filterName]
+            .map((filterValue) => `${filterName}.${filterValue}`)
+            .join(',');
+        })
+        .filter(filter => filter)
+        .join(',');
+
+      this.props.mutator.query.update({ filters });
+    });
+  }
+
+  handleCreate = (license) => {
     const { mutator } = this.props;
 
     mutator.records.POST(license)
@@ -44,24 +83,44 @@ export default class Licenses extends React.Component {
       });
   };
 
+  getActiveFilters = () => {
+    const { query } = this.props.resources;
+
+    if (!query || !query.filters) return undefined;
+
+    return query.filters
+      .split(',')
+      .reduce((filterMap, currentFilter) => {
+        const [name, value] = currentFilter.split('.');
+
+        if (!Array.isArray(filterMap[name])) {
+          filterMap[name] = [];
+        }
+
+        filterMap[name].push(value);
+        return filterMap;
+      }, {});
+  }
+
+  renderFilters = (onChange) => {
+    return (
+      <LicenseFilters
+        activeFilters={this.getActiveFilters()}
+        onChange={onChange}
+        resources={this.props.resources}
+      />
+    );
+  }
+
   render() {
-    // II Copied from ../ui-users/src/Users.js - I have no idea which of these might be needed for other things, or where this list
-    // is defined, so leaving it here in full, along with the signpost to Users.js to try and help the next lost soul who finds themselves here.
-    const { onSelectRow } = this.props;
-
-    const path = '/licenses';
-    packageInfo.stripes.route = path;
-    packageInfo.stripes.home = path;
-
     return (
       <SearchAndSort
+        browseOnly={this.props.browseOnly}
         columnMapping={{
-          id: 'ID',
           name: 'Name',
           description: 'Description'
         }}
         columnWidths={{
-          id: 300,
           name: 300,
           description: 'auto',
         }}
@@ -71,11 +130,13 @@ export default class Licenses extends React.Component {
         key="licenses"
         newRecordPerms="module.licenses.enabled"
         objectName="title"
-        onCreate={this.create}
-        onSelectRow={onSelectRow}
+        onCreate={this.handleCreate}
+        onFilterChange={this.handleFilterChange}
+        onSelectRow={this.props.onSelectRow}
         packageInfo={packageInfo}
         parentMutator={this.props.mutator}
         parentResources={this.props.resources}
+        renderFilters={this.renderFilters}
         resultCountIncrement={INITIAL_RESULT_COUNT}
         showSingleResult
         viewRecordComponent={ViewLicense}