+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.0.0 IN PROGRESS
 * Upgrade to Stripes 2.0
 * Removed usage of lookbehinds in regexes (avail in 2.0.1)
+* Added License Status filter to search.
 
 ## 1.1.0
 First Official Release for Q4 Release - 2018-12-11